+++ CHANGELOG.md
@@ -23,6 +23,7 @@ as necessary. Empty sections will not end in the release notes.
   If a fixed cache size setting has been explicitly configured, consider to change it to the fraction based one.
 - Relative hashes are now supported in table references, thus allowing SQL queries to specify a relative hash
   in the `FROM` clause, e.g. `FROM table1@main#1234^1`.
+- BigTable backend: ability to disable telemetry (which is enabled by default).
 
 ### Deprecations